requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
anyio==4.12.1
APScheduler==3.11.2
async-timeout==5.0.1
attrs==25.4.0
beautifulsoup4==4.14.3
certifi==2026.1.4
exceptiongroup==1.3.1
frozenlist==1.8.0
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
multidict==6.7.1
propcache==0.4.1
psycopg2-binary==2.9.11
python-telegram-bot==22.6
pytz==2025.2
soupsieve==2.8.3
SQLAlchemy==2.0.46
typing_extensions==4.15.0
tzlocal==5.3.1
yarl==1.22.0
